Code review: Fix path traversal, deadlock, division by zero, resource leaks, and UB
Security:
- Fix path traversal bypass in AvatarManager (path-component-aware check)
- Validate AES key/cipher sizes and check EVP Final return values
- Port validation in chat cleaner server
- Fix gsasl memory leaks on error paths in load.cpp
- Add HOME/XDG_CONFIG_HOME fallback in cleanerconfig

Correctness:
- Fix deadlock in PlayerData::operator< (use boost::lock for dual mutex)
- Fix uninitialized iterator in LocalBoard::determinePlayerNeedToShowCards
- Fix wrong variable (raise vs bet) in localplayer riverEngine
- Guard division by zero in odds calculation and flopEngine3
- Add missing return in CardsValue::holeCardsClass (UB fix)
- Fix signed-to-unsigned conversion in LocalBoard::distributePot
- Fix infinite loop when ASN.1 message exceeds buffer in load.cpp
- Fix capsNumberToTrigger <= 0 regex match in CapsFloodCheck
- Fix signed/unsigned mismatch in chat cleaner timestamps
- Add null check for tb in guilog writeLog
- Fix eventFilter UB (cast before type check) in gametableimpl
- Fix QStringList initialized with nullptr in chattools
- Fix leading space typo in gametablestylereader fallback lookup
- Add unsigned underflow guards in SessionData timers
- Fix ServerException wrong constructor args in servergame
- Cache state pointer before unlock in ClientThread::SetState

Modernization:
- Use C++23 memset_explicit in CryptHelper::SecureClearMemory
- Add override to InternalServerCallback in serverlobbythread
- Add override to SingleAsyncDBQuery::Next
- Delete copy/move for SoundEvents (Rule of Five)
- Pass by const ref for SetPlayerLastGames across DB interface chain
- Use inline static for Quote in mysqlpp_compat header (ODR fix)
- Use memcpy instead of reinterpret_cast for unaligned buffer
- Add delete for ServerGame copy constructor
- Use shared_from_this in AsioSendBuffer async handler
- Handle SSL streams in ServerLobbyThread::CloseSession
- Ensure thread joins in AvatarManager and Thread destructors

+++ src/chatcleaner/messagefilter.h
@@ -73,7 +73,7 @@ private:
 	};
 
 	struct ClientKickInfos {
-		size_t lastKickTimestamp = 0;
+		long lastKickTimestamp = 0;
 		int kickNumber = 0;
 	};
